# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virtigia Map Editor API",
        "description": "Token do autoryzacji wygenerujesz w panelu użytkownika: Profil -> Tokeny API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/npcs": {
            "get": {
                "tags": [
                    "NPC"
                ],
                "summary": "Pobranie listy NPC",
                "operationId": "getNpcsList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    },
                    {
                        "name": "base_npc_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginowana lista NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiNpcListCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "NPC"
                ],
                "summary": "Dodanie nowego NPC z pierwszą lokalizacją",
                "operationId": "storeNpc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "base_npc_id",
                                    "map_id",
                                    "x",
                                    "y"
                                ],
                                "properties": {
                                    "base_npc_id": {
                                        "type": "integer"
                                    },
                                    "map_id": {
                                        "type": "integer"
                                    },
                                    "x": {
                                        "type": "integer"
                                    },
                                    "y": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "NPC został dodany",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiNpcDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
        "/api/v1/npcs/{npcId}": {
            "get": {
                "tags": [
                    "NPC"
                ],
                "summary": "Pobranie pojedynczego NPC",
                "operationId": "getNpcById",
                "parameters": [
                    {
                        "name": "npcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiNpcDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "NPC nie został znaleziony"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "NPC"
                ],
                "summary": "Usunięcie NPC",
                "operationId": "destroyNpc",
                "parameters": [
                    {
                        "name": "npcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NPC został usunięty"
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "NPC nie został znaleziony"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "NPC"
                ],
                "summary": "Edycja NPC (name, lvl, rank na powiązanym base NPC)",
                "operationId": "updateNpc",
                "parameters": [
                    {
                        "name": "npcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "lvl": {
                                        "type": "integer"
                                    },
                                    "rank": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "NPC został zaktualizowany",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiNpcDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "NPC nie został znaleziony"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
        "/api/v1/npcs/{npcId}/locations": {
            "get": {
                "tags": [
                    "NPC Locations"
                ],
                "summary": "Pobranie paginowanej listy lokalizacji NPC",
                "operationId": "getNpcLocationsList",
                "parameters": [
                    {
                        "name": "npcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginowana lista lokalizacji NPC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiNpcLocationListCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "NPC nie został znaleziony"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "NPC Locations"
                ],
                "summary": "Dodanie nowej lokalizacji NPC",
                "operationId": "storeNpcLocation",
                "parameters": [
                    {
                        "name": "npcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "map_id",
                                    "x",
                                    "y"
                                ],
                                "properties": {
                                    "map_id": {
                                        "type": "integer"
                                    },
                                    "x": {
                                        "type": "integer"
                                    },
                                    "y": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lokalizacja została dodana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/App\\Http\\Resources\\ApiNpcLocationResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "NPC nie został znaleziony"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
        "/api/v1/npcs/{npcId}/locations/{locationId}": {
            "delete": {
                "tags": [
                    "NPC Locations"
                ],
                "summary": "Usunięcie lokalizacji NPC",
                "operationId": "destroyNpcLocation",
                "parameters": [
                    {
                        "name": "npcId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "locationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lokalizacja została usunięta"
                    },
                    "401": {
                        "description": "Brak lub nieprawidłowy token"
                    },
                    "404": {
                        "description": "NPC lub lokalizacja nie zostały znalezione"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "worldHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "API Token",
                "description": "Token do autoryzacji wygenerujesz w panelu użytkownika: Profil -> Tokeny API."
            },
            "worldHeader": {
                "type": "apiKey",
                "description": "Wymagany nagłówek świata. Dostępne wartości: retro, classic, legacy, demo.",
                "name": "X-World",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": [],
            "worldHeader": []
        }
    ]
}